pyproject.toml
[project]
name = "parakeet-clone"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "faster-whisper>=1.2.1",
    "openai>=1.54.0",
    "keyboard>=0.13.5",
    "numpy>=2.4.4",
    "pyaudiowpatch>=0.2.12.8",
    "python-dotenv>=1.2.2",
    "pymupdf>=1.24.0",
    "pyqt6>=6.7.0",
    "scipy>=1.17.1",
    "assemblyai>=0.40.0",
]

[project.scripts]
parakeet = "main:main"
